+++ R/utilities.R
@@ -355,3 +355,36 @@ vec_char_redu <- function(vec, Nchar=50){
     }, FUN.VALUE=character(1))
     return(res)
 }
+
+#' Show Reduced Targets
+#' 
+#' Reduce number of targets for each element of a character vector by 
+#' replacting the targets that beyond Ntar to '...'.  
+#' @param vec character vector, each element composed by a list of targets 
+#' symbols separated by '; '
+#' @param Ntar integer, for each element in the vec, number of targets to show
+#' @return character vector after reducing
+#' @examples 
+#' vec <- c("t1; t2; t3; t4; t5; t6", "t7; t8")
+#' vec2 <- tarReduce(vec, Ntar=5)
+#' @export
+#' 
+tarReduce <- function(vec, Ntar=5){
+    singleTarShot <- function(s, Ntar){
+        temp <- strsplit(s, ';\\s?')[[1]]
+        if(length(temp)>Ntar){
+            return(paste0(paste(temp[seq_len(Ntar)], collapse = "; "), "; ..."))
+        } else {
+            return(s)
+        }
+    }
+    return(sapply(vec, singleTarShot, Ntar, USE.NAMES = FALSE))
+}
+
+load_OrgDb <- function(OrgDb){
+    if (is(OrgDb, "character")) {
+        require(OrgDb, character.only = TRUE)
+        OrgDb <- eval(parse(text = OrgDb))
+    }
+    return(OrgDb)
+}